renamed the database_connector to data_connector and DatabaseConnector to DataConnector
