+++ ai_text_to_sql/data_connectors/mssql_connector.py
@@ -3,10 +3,10 @@ from typing import Dict, Text
 
 from sqlalchemy import create_engine
 
-from .database_connector import DatabaseConnector
+from .data_connector import DataConnector
 
 
-class MSSQLConnector(DatabaseConnector):
+class MSSQLConnector(DataConnector):
     """
     The Connector class for MSSQL databases.
