+++ ai_text_to_sql/data_connectors/sqlite_connector.py
@@ -2,10 +2,10 @@ from typing import Dict, Text
 
 from sqlalchemy import create_engine
 
-from .database_connector import DatabaseConnector
+from .data_connector import DataConnector
 
 
-class SQLiteConnector(DatabaseConnector):
+class SQLiteConnector(DataConnector):
     """
     The Connector class for SQLite databases.
