--- ai_text_to_sql/database_connectors/__init__.py
@@ -1,4 +1,0 @@
-from ai_text_to_sql.database_connectors.sqlite_connector import SQLiteConnector
-from ai_text_to_sql.database_connectors.mssql_connector import MSSQLConnector
-from ai_text_to_sql.database_connectors.mysql_connector import MySQLConnector
-from ai_text_to_sql.database_connectors.postgresql_connector import PostgreSQLConnector
